refactor to correctly expose API
