+++ nodl_python/nodl/warning.py
@@ -1,0 +1,4 @@
+class NoNodeInterfaceWarning(UserWarning):
+    """Warning raised when not marked as either server/client or pub/sub."""
+
+    pass
